Committed
I like the way the previous report and domain was formatted, so if you
guys can just format the Intermediate Status Report 3 that way, it would
be awesome. Also, the ISR can't implement code coverage until we
completely finish all implementations of the code for the next MileStone
3. Based on the previous Domain Model, I going to append this Domain
Model question onto the previous Domain Model.